dinov2_mobile/requirements.txt
# internvl/requirements.txt

torch>=2.0.0
transformers>=4.35.0
accelerate>=0.20.0
peft>=0.6.0
datasets>=2.14.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
huggingface_hub>=0.17.0
safetensors>=0.3.0
bitsandbytes>=0.41.0
scipy>=1.11.0
scikit-learn>=1.3.0
onnx-tf>=1.6.0
tf_keras>=2.19.0
onnx_graphsurgeon>=0.5.8
ai_edge_litert>=1.3.0
sng4onnx>=1.0.4

# Mobile deployment
onnx>=1.14.0
onnxruntime>=1.15.0
coremltools>=7.0
tensorflow>=2.13.0

# Development and testing
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.7.0
flake8>=6.0.0

